LayoutTests/media/media-source/media-managedmse-adts-aac.html

<!DOCTYPE html> <!-- webkit-test-runner [ ManagedMediaSourceEnabled=true MediaSourceEnabled=true ] -->
<html>
<head>
    <title>ManagedMediaSource ADTS-framed AAC</title>
    <script src="../video-test.js"></script>
    <script src="../utilities.js"></script>
    <script>
    // Companion to media-managedmse-aac-shared-format.html using an ADTS-framed
    // AAC stream as the MSE source.
    var source;
    var sourceBuffer;
    var buffer;

    window.addEventListener('load', async event => {
        try {
            findMediaElement();

            if (!ManagedMediaSource.isTypeSupported('audio/aac')) {
                consoleWrite('audio/aac not supported by ManagedMediaSource on this platform; skipping.');
                endTest();
                return;
            }

            run('video.disableRemotePlayback = true');
            source = new ManagedMediaSource();
            run('video.src = URL.createObjectURL(source)');
            await waitFor(source, 'sourceopen');
            waitFor(video, 'error').then(failTest);

            run('sourceBuffer = source.addSourceBuffer(\'audio/aac\')');

            const response = await fetch('./content/test-adts.aac');
            buffer = await response.arrayBuffer();
            run('sourceBuffer.appendBuffer(buffer)');

            await waitFor(sourceBuffer, 'updateend');
            testExpected('sourceBuffer.buffered.length', '1');
            consoleWrite(`sourceBuffer.buffered = ${ timeRangesToString(sourceBuffer.buffered) }`);

            run('source.endOfStream()');
            await waitFor(source, 'sourceended');

            testExpected('video.readyState', '4'); // HAVE_ENOUGH_DATA

            run('video.play()');
            await waitFor(video, 'playing');

            await waitForConditionOrTimeout('video.currentTime > 0', false, 1000);
            testExpected('video.paused', false);

            run('video.pause()');
            endTest();
        } catch (e) {
            failTest(`Caught exception: "${e}"`);
        }
    });
    </script>
</head>
<body>
    <video controls disableremoteplayback></video>
</body>
</html>
